requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.7
django-extra-views==0.13.0
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
